# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Employee API
  version: 1.0.0
  description: |
    Main API for managing employees. 
    For complex data queries, use our **[GraphQL Playground](http://localhost:4000)**.
servers:
  - url: http://localhost:3000

paths:
  /employees:
    get:
      summary: Get all employees
      description: Returns the list of all employees
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
    post:
      summary: Create a new employee
      description: Adds a new employee to the mock database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        201:
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        400:
          description: Invalid input
  /seed:
    post:
      summary: Create a new employee
      description: Adds a new employee to the mock database.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                count:
                  type: integer
                  example: 20
                department:
                  type: string
                  example: "Engineering" # Optional field
      responses:
        200:
          description: Successfully seeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  count: { type: integer }
  /emp_response:
    post:
      summary: Set response to employee get method
      description: Set response to employee get method.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: integer
                  example: 200
                message:
                  type: string
                  example: "This is the error message from server" # Optional field
      responses:
        200:
          description: Successfully changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
  /graphql:
    post:
      summary: Execute a GraphQL query with renamed users
      description: |
        Forwards a GraphQL query to the internal server, intercepts the response, 
        and prefixes all usernames with 'trik_'.
      tags:
        - Mocking GraphQL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: The GraphQL query string.
                  example: "query { employees { user { username } } }"
                variables:
                  type: object
                  description: Optional variables for the GraphQL query.
                  example: {}
      responses:
        '200':
          description: Successfully transformed employee data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphQLResponse'
        '400':
          description: Bad Request (Invalid GraphQL syntax)
        '500':
          description: Internal Server Error
                  
   

components:
  schemas:
    GraphQLResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            employees:
              type: array
              items:
                $ref: '#/components/schemas/Employee'
    User:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string }
        createdAt: { type: string }
    Employee:
      type: object
      properties:
        id: { type: string }
        user: { $ref: '#/components/schemas/User' }
        department:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
        bossId: { type: string }
        subordinates:
          type: array
          items: { type: string }
    EmpResponse:
      type: object
      required:
        - status
      properties:
        status: { type: number}
        message: { type: string }
    Department:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: "Marketing"